Move commands to commands.py, add logging, cleaned Dockerfile
commands.py
-----------
Move functions relating to bot commands (i.e. /whodabest) to separate file

logging
-------
Add logging for the above function. Log messages are different for group chats
versus private chats. Note that the check =='' or ==None is for back-
compatibility with older versions of python-telegram-bot (not 100% sure, but
one worked on rpi and the other in docker, so that's my best guess)

Dockerfile
----------
Remove instruction to copy requirements.txt. The file itself was removed
previously, but the instruction was not. Would cause errors in building the
docker image. Also cleaned the copy keys.py soyaya.py into *.py, which will be
helpful with additional (if) .pys.