Remove cli, add /set_me, /cancel, /about commands
listen.py
---------
Add set_me for the set_me command, set_me_message to handle messages
from users after the set_me command, cancel to exit the set_me 'state',
and an /about message.

utils.py
--------
Create a custom filter class to separate regular messages to the bot to
those that are meant as a reply to /set_me

__main__.py
-----------
Remove the cli portion of soyaya_bot, since there already seems to be
a built-in JobQueue class to handle periodic tasks (previously this was
to be the scope of look).

cli.py
------
Remove the cli class to reflect changes in __main__.py (cli.py now
redundant)