Add new restrict_to_registered for set_me
restrict_to_registered is a new decorator for set_me.

Previously, when an unregistered user uses /set_me in an allowed group,
bot would respond with for user to pm the bot, even though the user is
not yet registered.

Now, set_me is implemented with nested decorators such that

(1) if user is unregistered, direct the user to register, _THEN_
(2) if user uses set_me outside of a pm, direct the user to pm the bot